pyproject.toml
[project]
name = "chrome-extension-builder"
version = "0.1.0"
description = "A Replit-like website for building Chrome extensions through chat interface"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "selenium>=4.15.0",
    "playwright>=1.40.0",
    "google-generativeai>=0.3.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto" 
